// web_ui/src/pages/project-details/components/project-test/media-items-bucket.interface.ts
// INTEL CONFIDENTIAL
//
// Copyright (C) 2024 Intel Corporation
//
// This software and the related documents are Intel copyrighted materials, and your use of them is governed by
// the express license under which they were provided to you ("License"). Unless the License provides otherwise,
// you may not use, modify, copy, publish, distribute, disclose or transmit this software or the related documents
// without Intel's prior written permission.
//
// This software and the related documents are provided as is, with no express or implied warranties,
// other than those that are expressly stated in the License.

export enum MediaItemsBucketType {
    ABOVE_THRESHOLD = 'ABOVE_THRESHOLD',
    BELOW_THRESHOLD = 'BELOW_THRESHOLD',
}

export enum MediaItemsBucketTitle {
    ABOVE_THRESHOLD = 'Above threshold',
    BELOW_THRESHOLD = 'Below threshold',
    CORRECT = 'Correct',
    INCORRECT = 'Incorrect',
}